+++ src/main/java/Phonesonal/PhoneBE/repository/RecommendMealRepository.java
@@ -1,4 +1,4 @@
-package Phonesonal.PhoneBE.repository.Food;
+package Phonesonal.PhoneBE.repository;
 import Phonesonal.PhoneBE.domain.RecommendMeal;
 
 import Phonesonal.PhoneBE.domain.common.GoalPeriod;
