+++ src/main/java/Phonesonal/PhoneBE/service/Food/FoodQueryServiceImpl.java
@@ -1,7 +1,7 @@
-package Phonesonal.PhoneBE.service;
+package Phonesonal.PhoneBE.service.Food;
 
 import Phonesonal.PhoneBE.domain.Food;
-import Phonesonal.PhoneBE.repository.Food.FoodRepository;
+import Phonesonal.PhoneBE.repository.FoodRepository;
 import Phonesonal.PhoneBE.web.dto.Food.SearchFoodResponseDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
