+++ src/main/java/Phonesonal/PhoneBE/service/Food/UserMealCommandServiceImpl.java
@@ -1,14 +1,13 @@
-package Phonesonal.PhoneBE.service;
+package Phonesonal.PhoneBE.service.Food;
 
 import Phonesonal.PhoneBE.apiPayload.code.util.DateUtil;
 import Phonesonal.PhoneBE.domain.Food;
 import Phonesonal.PhoneBE.domain.User;
 import Phonesonal.PhoneBE.domain.UserMeal;
 import Phonesonal.PhoneBE.domain.common.GoalPeriod;
-import Phonesonal.PhoneBE.domain.enums.MealTime;
-import Phonesonal.PhoneBE.repository.Food.FoodRepository;
+import Phonesonal.PhoneBE.repository.FoodRepository;
 import Phonesonal.PhoneBE.repository.GoalPeriodRepository;
-import Phonesonal.PhoneBE.repository.Food.UserMealRepository;
+import Phonesonal.PhoneBE.repository.UserMealRepository;
 import Phonesonal.PhoneBE.repository.UserRepository;
 import Phonesonal.PhoneBE.web.dto.Food.AddUserMealCustomRequestDTO;
 import Phonesonal.PhoneBE.web.dto.Food.AddUserMealFromFoodRequestDTO;
