+++ src/main/java/Phonesonal/PhoneBE/service/RecommendMealService/RecommendMealQueryServiceImpl.java
@@ -5,8 +5,8 @@ import Phonesonal.PhoneBE.domain.RecommendMeal;
 import Phonesonal.PhoneBE.domain.common.GoalPeriod;
 import Phonesonal.PhoneBE.domain.enums.MealTime;
 import Phonesonal.PhoneBE.repository.GoalPeriodRepository;
-import Phonesonal.PhoneBE.repository.Food.RecommendMealRepository;
-import Phonesonal.PhoneBE.web.dto.RecommendMealResponseDTO;
+import Phonesonal.PhoneBE.repository.RecommendMealRepository;
+import Phonesonal.PhoneBE.web.dto.Food.RecommendMealResponseDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
