+++ src/main/java/Phonesonal/PhoneBE/web/controller/UserMealController.java
@@ -4,21 +4,18 @@ import Phonesonal.PhoneBE.apiPayload.ApiResponse;
 import Phonesonal.PhoneBE.apiPayload.code.status.SuccessStatus;
 import Phonesonal.PhoneBE.domain.enums.MealTime;
 import Phonesonal.PhoneBE.security.CustomUserDetails;
-import Phonesonal.PhoneBE.service.UserMealCommandService;
-import Phonesonal.PhoneBE.service.UserMealQueryService;
-import Phonesonal.PhoneBE.web.dto.CompleteStatusResponseDTO;
+import Phonesonal.PhoneBE.service.Food.UserMealCommandService;
+import Phonesonal.PhoneBE.service.Food.UserMealQueryService;
 import Phonesonal.PhoneBE.web.dto.Food.AddUserMealCustomRequestDTO;
 import Phonesonal.PhoneBE.web.dto.Food.AddUserMealFromFoodRequestDTO;
 import Phonesonal.PhoneBE.web.dto.Food.UserMealResponseDTO;
-import Phonesonal.PhoneBE.web.dto.RecommendMealRequestDTO;
-import Phonesonal.PhoneBE.web.dto.UpdateUserMealQuantityRequestDTO;
+import Phonesonal.PhoneBE.web.dto.Food.UpdateUserMealQuantityRequestDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import Phonesonal.PhoneBE.web.dto.Food.AddUserMealFromFoodRequestDTO;
 
 import java.time.LocalDate;
 import java.util.List;
