+++ src/main/java/Phonesonal/PhoneBE/web/dto/Food/RecommendMealRequestDTO.java
@@ -1,11 +1,10 @@
-package Phonesonal.PhoneBE.web.dto;
+package Phonesonal.PhoneBE.web.dto.Food;
 
 import Phonesonal.PhoneBE.domain.enums.MealTime;
 import lombok.*;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import java.time.LocalDate;
-import java.util.List;
 
 public class RecommendMealRequestDTO {
