+++ src/main/java/Phonesonal/PhoneBE/web/dto/Food/UpdateCompleteStatusRequestDTO.java
@@ -1,4 +1,4 @@
-package Phonesonal.PhoneBE.web.dto;
+package Phonesonal.PhoneBE.web.dto.Food;
 
 import Phonesonal.PhoneBE.domain.enums.MealTime;
 import lombok.*;
